refactor[#101] : unAttached, attached...Container -> 하나의 ReferenceListWidget 으로 리팩토링
