+++ extension/src/features/reference/ui/index.ts
@@ -4,5 +4,3 @@ export * from "./ResetReferenceButton";
 export * from "./CopyReferenceListButton";
 export * from "./AutoConvertingToggle";
 export * from "./ConvertToReferenceButton";
-export * from "./AttachedReferenceList";
-export * from "./UnAttachedReferenceList";
