+++ extension/src/widgets/reference/ui/index.ts
@@ -1,2 +1,1 @@
-export * from "./AttachedReferenceContainer";
-export * from "./UnAttachedReferenceContainer";
+export * from "./ReferenceListWidget";
